ActualizacionModuloCuentas
Actualización de los servicios y los controladores

+++ cuentas/src/main/java/com/banquito/core/banking/cuentas/domain/Transaccion.java
@@ -1,19 +1,9 @@
 package com.banquito.core.banking.cuentas.domain;
 
-import java.math.BigDecimal;
-import java.security.Timestamp;
+import java.sql.Timestamp;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-import jakarta.persistence.Version;
+
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -48,10 +38,10 @@ public class Transaccion {
     private String tipoAfectacion;
 
     @Column(name = "VALOR_DEBE", nullable = false, precision = 18, scale = 2)
-    private BigDecimal valorDebe;
+    private Long valorDebe;
 
     @Column(name = "VALOR_HABER", nullable = false, precision = 18, scale = 2)
-    private BigDecimal valorHaber;
+    private Long valorHaber;
 
     @Column(name = "TIPO_TRANSACCION", nullable = false, length = 8)
     private String tipoTransaccion;
